old/inference/requirements.txt
requests
python-dotenv
tiktoken
marvin
pydantic